refactor!: Split TrackStateProxy into own header, public  (#2807)
This moves TrackStateProxy into its own header file, and moves it out
of the `detail_lt` namespace.

Blocked by: 
- https://github.com/acts-project/acts/pull/2712

BREAKING CHANGE: `Acts::detail_lt::TrackStateProxy` moves to `Acts::TrackStateProxy` and include `Acts/EventData/TrackStateProxy.hpp`
